Avoid public type collisions in GlobalTypeRewriter
Move BrandTypeIterator from MinimizeRecGroups to wasm-type-shape.h and
use it in a new UniqueRecGroups utility that can rebuild types to be
distinct from previously seen rec groups. Use UniqueRecGroups in
GlobalTypeRewriter to ensure the newly built private types do not
conflict with public types.

Split off from #8119 because this can land sooner.
