+++ src/ir/type-updating.cpp
@@ -26,7 +26,28 @@
 
 namespace wasm {
 
-GlobalTypeRewriter::GlobalTypeRewriter(Module& wasm) : wasm(wasm) {}
+GlobalTypeRewriter::GlobalTypeRewriter(Module& wasm)
+  : wasm(wasm), publicGroups(wasm.features) {
+  // Find the heap types that are not publicly observable. Even in a closed
+  // world scenario, don't modify public types because we assume that they may
+  // be reflected on or used for linking. Figure out where each private type
+  // will be located in the builder.
+  typeInfo = ModuleUtils::collectHeapTypeInfo(
+    wasm,
+    ModuleUtils::TypeInclusion::UsedIRTypes,
+    ModuleUtils::VisibilityHandling::FindVisibility);
+
+  std::unordered_set<RecGroup> seenGroups;
+  for (auto& [type, info] : typeInfo) {
+    if (info.visibility == ModuleUtils::Visibility::Public) {
+      auto group = type.getRecGroup();
+      if (seenGroups.insert(type.getRecGroup()).second) {
+        std::vector<HeapType> groupTypes(group.begin(), group.end());
+        publicGroups.insert(std::move(groupTypes));
+      }
+    }
+  }
+}
 
 void GlobalTypeRewriter::update() {
   mapTypes(rebuildTypes(getSortedTypes(getPrivatePredecessors())));
@@ -34,16 +55,8 @@ void GlobalTypeRewriter::update() {
 
 GlobalTypeRewriter::PredecessorGraph
 GlobalTypeRewriter::getPrivatePredecessors() {
-  // Find the heap types that are not publicly observable. Even in a closed
-  // world scenario, don't modify public types because we assume that they may
-  // be reflected on or used for linking. Figure out where each private type
-  // will be located in the builder.
-  auto typeInfo = ModuleUtils::collectHeapTypeInfo(
-    wasm,
-    ModuleUtils::TypeInclusion::UsedIRTypes,
-    ModuleUtils::VisibilityHandling::FindVisibility);
-
-  // Check if a type is private, by looking up its info.
+  // Check if a type is private, looking for its info (if there is none, it is
+  // not private).
   auto isPublic = [&](HeapType type) {
     auto it = typeInfo.find(type);
     assert(it != typeInfo.end());
@@ -185,11 +198,8 @@ GlobalTypeRewriter::rebuildTypes(std::vector<HeapType> types) {
             << " at index " << err->index;
   }
 #endif
-  auto& newTypes = *buildResults;
-
-  // TODO: It is possible that the newly built rec group matches some public rec
-  // group. If that is the case, we need to try a different permutation of the
-  // types or add a brand type to distinguish the private types.
+  // Ensure the new types are different from any public rec group.
+  const auto& newTypes = publicGroups.insert(*buildResults);
 
   // Map the old types to the new ones.
   TypeMap oldToNewTypes;
