+++ src/passes/MinimizeRecGroups.cpp
@@ -100,84 +100,6 @@ struct TypeSCCs
   }
 };
 
-// After all their permutations with distinct shapes have been used, different
-// groups with the same shapes must be differentiated by adding in a "brand"
-// type. Even with a brand mixed in, we might run out of permutations with
-// distinct shapes, in which case we need a new brand type. This iterator
-// provides an infinite sequence of possible brand types, prioritizing those
-// with the most compact encoding.
-struct BrandTypeIterator {
-  static constexpr Index optionCount = 18;
-  static constexpr std::array<Field, optionCount> fieldOptions = {{
-    Field(Field::i8, Mutable),
-    Field(Field::i16, Mutable),
-    Field(Type::i32, Mutable),
-    Field(Type::i64, Mutable),
-    Field(Type::f32, Mutable),
-    Field(Type::f64, Mutable),
-    Field(Type(HeapType::any, Nullable), Mutable),
-    Field(Type(HeapType::func, Nullable), Mutable),
-    Field(Type(HeapType::ext, Nullable), Mutable),
-    Field(Type(HeapType::none, Nullable), Mutable),
-    Field(Type(HeapType::nofunc, Nullable), Mutable),
-    Field(Type(HeapType::noext, Nullable), Mutable),
-    Field(Type(HeapType::any, NonNullable), Mutable),
-    Field(Type(HeapType::func, NonNullable), Mutable),
-    Field(Type(HeapType::ext, NonNullable), Mutable),
-    Field(Type(HeapType::none, NonNullable), Mutable),
-    Field(Type(HeapType::nofunc, NonNullable), Mutable),
-    Field(Type(HeapType::noext, NonNullable), Mutable),
-  }};
-
-  struct FieldInfo {
-    uint8_t index = 0;
-    bool immutable = false;
-
-    operator Field() const {
-      auto field = fieldOptions[index];
-      if (immutable) {
-        field.mutable_ = Immutable;
-      }
-      return field;
-    }
-
-    bool advance() {
-      if (!immutable) {
-        immutable = true;
-        return true;
-      }
-      immutable = false;
-      index = (index + 1) % optionCount;
-      return index != 0;
-    }
-  };
-
-  bool useArray = false;
-  std::vector<FieldInfo> fields;
-
-  HeapType operator*() const {
-    if (useArray) {
-      return Array(fields[0]);
-    }
-    return Struct(std::vector<Field>(fields.begin(), fields.end()));
-  }
-
-  BrandTypeIterator& operator++() {
-    for (Index i = fields.size(); i > 0; --i) {
-      if (fields[i - 1].advance()) {
-        return *this;
-      }
-    }
-    if (useArray) {
-      useArray = false;
-      return *this;
-    }
-    fields.emplace_back();
-    useArray = fields.size() == 1;
-    return *this;
-  }
-};
-
 // Create an adjacency list with edges from supertype to subtype and from
 // described type to descriptor.
 std::vector<std::vector<Index>>
